+++ src/BugsnagServiceProvider.php
@@ -190,6 +190,7 @@ class BugsnagServiceProvider extends ServiceProvider
 
             $client->setReleaseStage(isset($config['release_stage']) ? $config['release_stage'] : $app->environment());
             $client->setHostname(isset($config['hostname']) ? $config['hostname'] : null);
+            $client->getConfig()->mergeDeviceData(['runtimeVersions' => $this->getRuntimeVersion()]);
 
             $client->setFallbackType($app->runningInConsole() ? 'Console' : 'HTTP');
             $client->setAppType(isset($config['app_type']) ? $config['app_type'] : null);
@@ -398,6 +399,25 @@ class BugsnagServiceProvider extends ServiceProvider
         };
 
         $sessionTracker->setStorageFunction($genericStorage);
+    }
+
+    /**
+     * Returns the framework name and version to add to the device data.
+     *
+     * Attempt to parse a semantic framework version from $app or else return
+     * the full version string.
+     * e.g. Lumen: "Lumen (x.x.x) (Laravel Components y.y.*)" => "x.x.x"
+     *
+     * @return array
+     */
+    protected function getRuntimeVersion()
+    {
+        $version = $this->app->version();
+        if (preg_match('/(\d+\.\d+\.\d+)/', $version, $versionMatches)) {
+            $version = $versionMatches[0];
+        }
+        return [ ($this->app instanceof LumenApplication ? 'lumen' : 'laravel' ) => $version ];
+
     }
 
     /**