+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Changelog
   [#344](https://github.com/bugsnag/bugsnag-laravel/pull/344)
   [Mozammil Khodabacchas](https://github.com/mozammil)
 
+## TBD
+
+* Add Laravel/Lumen version string to report and session payloads (device.runtimeVersions)
+  [#352](https://github.com/bugsnag/bugsnag-laravel/pull/352)
+
 ## 2.15.2 (2019-01-23)
 
 ### Fixes